backend/django_project/requirements/requirements-production.txt
Django>=4.0
jinja2
psycopg2-binary
requests
daphne
pymemcache
gunicorn
django-jazzmin
django-cors-headers
ipython
uritemplate
reportlab
gunicorn
celery[redis]
autobahn
Twisted[tls,http2]
redis
Pillow
firebase_admin
djangorestframework-xml
djangorestframework
channels
channels_redis
python-slugify
beautifulsoup4
html5lib
django_pandas
numpy
qrcode
xlsxwriter
django-filter
aiofiles